Refactor: router
