+++ example/nodejs/src/controllers/ping.ts
@@ -1,0 +1,6 @@
+import { RouteContext } from '../application';
+
+// Ping
+export const getPing = async (context: RouteContext): Promise<void> => {
+  context.res.setBody('pong');
+};
