+++ example/nodejs/src/routes/auditlogs.ts
@@ -1,11 +1,9 @@
-import { domainsAuditLog } from '@viron/lib';
-import { RouteContext } from '../application';
+import { domainsOas } from '@viron/lib';
+import * as controllers from '../controllers/auditlogs';
+import { Route } from '../router';
 
-// 監査ログ一覧
-export const listVironAuditlogs = async (
-  context: RouteContext
-): Promise<void> => {
-  const { size, page, sort, ...conditions } = context.params.query;
-  const result = await domainsAuditLog.list(conditions, size, page, sort);
-  context.res.json(result);
+export const routeAuditlogs: Route = {
+  name: 'auditlogs',
+  oasPath: domainsOas.getPath('auditlogs'),
+  controllers,
 };
