+++ example/nodejs/src/routes/auth.ts
@@ -1,72 +1,9 @@
-import {
-  domainsAuth,
-  genAuthorizationCookie,
-  genOAuthStateCookie,
-  mismatchState,
-  COOKIE_KEY,
-  HTTP_HEADER,
-} from '@viron/lib';
-import { RouteContext } from '../application';
-import { ctx } from '../context';
+import { domainsOas } from '@viron/lib';
+import * as controllers from '../controllers/auth';
+import { Route } from '../router';
 
-// サインアウト
-export const signout = async (context: RouteContext): Promise<void> => {
-  const token = context.req.cookies[COOKIE_KEY.VIRON_AUTHORIZATION];
-  context.origRes.clearCookie(COOKIE_KEY.VIRON_AUTHORIZATION);
-  await domainsAuth.signout(token);
-  context.res.status(204).end();
-};
-
-// Emailサインイン
-export const signinEmail = async (context: RouteContext): Promise<void> => {
-  const { email, password } = context.requestBody;
-  const token = await domainsAuth.signinEmail(email, password);
-  context.res.setHeader(
-    HTTP_HEADER.SET_COOKIE,
-    genAuthorizationCookie(token, { maxAge: ctx.config.auth.jwt.expirationSec })
-  );
-  context.res.status(204).end();
-};
-
-// GoogleOAuth2の認可画面へリダイレクト
-export const oauth2GoogleAuthorization = async (
-  context: RouteContext
-): Promise<void> => {
-  const { redirectUri } = context.params.query;
-  const state = domainsAuth.genState();
-
-  const authorizationUrl = domainsAuth.getGoogleOAuth2AuthorizationUrl(
-    redirectUri,
-    state,
-    ctx.config.auth.googleOAuth2
-  );
-
-  context.res.setHeader(HTTP_HEADER.SET_COOKIE, genOAuthStateCookie(state));
-  context.res.setHeader(HTTP_HEADER.LOCATION, authorizationUrl);
-  context.res.status(301).end();
-};
-
-// GoogleOAuth2のコールバック
-export const oauth2GoogleCallback = async (
-  context: RouteContext
-): Promise<void> => {
-  const cookieState = context.req.cookies[COOKIE_KEY.OAUTH2_STATE];
-  const { code, state, redirectUri } = context.requestBody;
-
-  if (!cookieState || !state || cookieState !== state) {
-    throw mismatchState();
-  }
-
-  const token = await domainsAuth.signinGoogleOAuth2(
-    code,
-    redirectUri,
-    ctx.config.auth.googleOAuth2
-  );
-  context.res.setHeader(
-    HTTP_HEADER.SET_COOKIE,
-    genAuthorizationCookie(token, {
-      maxAge: ctx.config.auth.jwt.expirationSec,
-    })
-  );
-  context.res.status(204).end();
+export const routeAuth: Route = {
+  name: 'auth',
+  oasPath: domainsOas.getPath('auth'),
+  controllers,
 };
