+++ example/nodejs/src/routes/resources.ts
@@ -1,19 +1,9 @@
-import { HTTP_HEADER } from '@viron/lib';
-import { contentType } from 'mime-types';
-import { RouteContext } from '../application';
-import { exportResources } from '../domains/resources';
+import * as controllers from '../controllers/resources';
+import { oasPath } from '../domains/oas';
+import { Route } from '../router';
 
-// データダウンロード
-export const downloadResources = async (
-  context: RouteContext
-): Promise<void> => {
-  const { resourceName } = context.params.path;
-  const { format = 'json' } = context.params.query;
-  const result = await exportResources(resourceName, format);
-  context.res.header(HTTP_HEADER.CONTENT_TYPE, contentType(format) || '');
-  context.res.header(
-    HTTP_HEADER.CONTENT_DISPOSITION,
-    `attachment; filename="${resourceName}-${Date.now()}.${format}"`
-  );
-  context.res.setBody(result).end();
+export const routeResources: Route = {
+  name: 'resources',
+  oasPath: oasPath('resources'),
+  controllers,
 };
