+++ example/nodejs/src/routes/users.ts
@@ -1,27 +1,9 @@
-import { RouteContext } from '../application';
-import { list, createOne, updateOneById, removeOneById } from '../domains/user';
+import * as controllers from '../controllers/users';
+import { oasPath } from '../domains/oas';
+import { Route } from '../router';
 
-// ユーザー一覧
-export const listUsers = async (context: RouteContext): Promise<void> => {
-  const { size, page, sort, ...conditions } = context.params.query;
-  const result = await list(conditions, size, page, sort);
-  context.res.json(result);
-};
-
-// ユーザー作成
-export const createUser = async (context: RouteContext): Promise<void> => {
-  const user = await createOne(context.requestBody);
-  context.res.status(201).json(user);
-};
-
-// ユーザー更新
-export const updateUser = async (context: RouteContext): Promise<void> => {
-  await updateOneById(context.params.path.userId, context.requestBody);
-  context.res.status(204).end();
-};
-
-// ユーザー削除
-export const removeUser = async (context: RouteContext): Promise<void> => {
-  await removeOneById(context.params.path.userId);
-  context.res.status(204).end();
+export const routeUsers: Route = {
+  name: 'users',
+  oasPath: oasPath('users'),
+  controllers,
 };
